+++ src/main/java/com/nf404/devshop/user/dao/impl/UserDAOImpl.java
@@ -1,8 +1,8 @@
-package com.nf404.devshop.dao.impl;
+package com.nf404.devshop.user.dao.impl;
 
-import com.nf404.devshop.dao.UserDAO;
-import com.nf404.devshop.mapper.UserMapper;
-import com.nf404.devshop.model.UserDTO;
+import com.nf404.devshop.user.dao.UserDAO;
+import com.nf404.devshop.user.mapper.UserMapper;
+import com.nf404.devshop.user.model.UserDTO;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
